rename the package from ui.helps to better name configs
