+++ src/main/java/com/google/jenkins/plugins/computeengine/InstanceConfiguration.java
@@ -40,12 +40,12 @@ import com.google.cloud.graphite.platforms.plugin.client.ComputeClient;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
 import com.google.jenkins.plugins.computeengine.client.ClientUtil;
+import com.google.jenkins.plugins.computeengine.configs.PreemptibleVm;
+import com.google.jenkins.plugins.computeengine.configs.ProvisioningType;
+import com.google.jenkins.plugins.computeengine.configs.Standard;
 import com.google.jenkins.plugins.computeengine.ssh.GoogleKeyCredential;
 import com.google.jenkins.plugins.computeengine.ssh.GoogleKeyPair;
 import com.google.jenkins.plugins.computeengine.ssh.GooglePrivateKey;
-import com.google.jenkins.plugins.computeengine.ui.helpers.PreemptibleVm;
-import com.google.jenkins.plugins.computeengine.ui.helpers.ProvisioningType;
-import com.google.jenkins.plugins.computeengine.ui.helpers.Standard;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import hudson.Extension;
