+++ src/main/resources/com/google/jenkins/plugins/computeengine/configs/ProvisioningType/config.jelly
