+++ src/test/java/com/google/jenkins/plugins/computeengine/ConfigAsCodeTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThrows;
 
-import com.google.jenkins.plugins.computeengine.ui.helpers.PreemptibleVm;
+import com.google.jenkins.plugins.computeengine.configs.PreemptibleVm;
 import com.google.jenkins.plugins.credentials.oauth.GoogleRobotPrivateKeyCredentials;
 import hudson.model.Node;
 import io.jenkins.plugins.casc.misc.ConfiguredWithCode;
