+++ src/test/java/com/google/jenkins/plugins/computeengine/InstanceConfigurationTest.java
@@ -36,8 +36,8 @@ import com.google.api.services.compute.model.Subnetwork;
 import com.google.api.services.compute.model.Zone;
 import com.google.cloud.graphite.platforms.plugin.client.ComputeClient;
 import com.google.common.collect.ImmutableList;
-import com.google.jenkins.plugins.computeengine.ui.helpers.PreemptibleVm;
-import com.google.jenkins.plugins.computeengine.ui.helpers.Standard;
+import com.google.jenkins.plugins.computeengine.configs.PreemptibleVm;
+import com.google.jenkins.plugins.computeengine.configs.Standard;
 import hudson.model.Node;
 import hudson.util.FormValidation;
 import java.util.ArrayList;
