+++ src/test/java/com/google/jenkins/plugins/computeengine/integration/ITUtil.java
@@ -59,10 +59,10 @@ import com.google.jenkins.plugins.computeengine.NetworkInterfaceSingleStack;
 import com.google.jenkins.plugins.computeengine.SshConfiguration;
 import com.google.jenkins.plugins.computeengine.WindowsConfiguration;
 import com.google.jenkins.plugins.computeengine.client.ClientUtil;
+import com.google.jenkins.plugins.computeengine.configs.PreemptibleVm;
+import com.google.jenkins.plugins.computeengine.configs.Standard;
 import com.google.jenkins.plugins.computeengine.ssh.GoogleKeyCredential;
 import com.google.jenkins.plugins.computeengine.ssh.GoogleKeyPair;
-import com.google.jenkins.plugins.computeengine.ui.helpers.PreemptibleVm;
-import com.google.jenkins.plugins.computeengine.ui.helpers.Standard;
 import com.google.jenkins.plugins.credentials.oauth.GoogleRobotPrivateKeyCredentials;
 import com.google.jenkins.plugins.credentials.oauth.JsonServiceAccountConfig;
 import hudson.model.Node;
