+++ src/test/resources/com/google/jenkins/plugins/computeengine/SchedulingTests/testNoMaxRunDurationPreemptible/config.xml
@@ -12,7 +12,7 @@
             <configurations>
                 <com.google.jenkins.plugins.computeengine.InstanceConfiguration>
                     <preemptible>false</preemptible>
-                    <provisioningType class="com.google.jenkins.plugins.computeengine.ui.helpers.PreemptibleVm">
+                    <provisioningType class="com.google.jenkins.plugins.computeengine.configs.PreemptibleVm">
                     </provisioningType>
                 </com.google.jenkins.plugins.computeengine.InstanceConfiguration>
             </configurations>
