+++ src/test/resources/com/google/jenkins/plugins/computeengine/SchedulingTests/testProvisioningTypeOverridesPreemptible/config.xml
@@ -13,7 +13,7 @@
             <configurations>
                 <com.google.jenkins.plugins.computeengine.InstanceConfiguration>
                     <preemptible>true</preemptible>
-                    <provisioningType class="com.google.jenkins.plugins.computeengine.ui.helpers.SpotVm">
+                    <provisioningType class="com.google.jenkins.plugins.computeengine.configs.SpotVm">
                     </provisioningType>
                 </com.google.jenkins.plugins.computeengine.InstanceConfiguration>
             </configurations>
